move decorator to common package
