+++ busy_beaver/apps/poller/api/upcoming_events.py
@@ -6,7 +6,7 @@ from flask.views import MethodView
 from busy_beaver.apps.upcoming_events.workflow import (
     post_upcoming_events_message_to_slack,
 )
-from busy_beaver.blueprints.decorators import admin_role_required
+from busy_beaver.common.decorators import admin_role_required
 from busy_beaver.toolbox import make_response
 
 logger = logging.getLogger(__name__)
