+++ busy_beaver/apps/poller/api/youtube_post.py
@@ -6,7 +6,7 @@ from sqlalchemy import desc
 
 from busy_beaver import config
 from busy_beaver.adapters.youtube import YouTubeAdapter
-from busy_beaver.blueprints.decorators import admin_role_required
+from busy_beaver.common.decorators import admin_role_required
 from busy_beaver.clients import chipy_slack
 from busy_beaver.toolbox import make_response
 from busy_beaver.extensions import db
