+++ busy_beaver/common/decorators.py
